cart apis added

+++ models/order.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Order struct {
-	ID          uint       `gorm:"primary_key" json:"id"`
-	UserID      uint       `gorm:"not null" json:"user_id"`
+	ID          uint64       `gorm:"primary_key" json:"id"`
+	UserID      uint64       `gorm:"not null" json:"user_id"`
 	Products    string     `gorm:"type:json;not null" json:"products"`
 	OrderStatus string     `gorm:"type:enum('placed', 'dispatched', 'completed', 'cancelled');not null" json:"order_status"`
 	Dispatched  time.Time  `gorm:"type:datetime" json:"dispatched"`
